requirements.txt
Flask
tensorflow
opencv-python-headless
Pillow
gunicorn
